# docs/swagger.yaml
definitions:
  document.EsProduct:
    properties:
      attrValueList:
        items:
          $ref: '#/definitions/domain.EsProductAttributeValue'
        type: array
      brandId:
        type: integer
      brandName:
        type: string
      id:
        description: model.PmsProduct
        type: integer
      keywords:
        type: string
      name:
        type: string
      newStatus:
        type: integer
      pic:
        type: string
      price:
        type: number
      productCategoryId:
        type: integer
      productCategoryName:
        type: string
      productSn:
        type: string
      promotionType:
        type: integer
      recommandStatus:
        type: integer
      sale:
        type: integer
      sort:
        type: integer
      stock:
        type: integer
      subTitle:
        type: string
    type: object
  domain.EsProductAttributeValue:
    properties:
      id:
        type: integer
      name:
        type: string
      productAttributeId:
        type: integer
      type:
        type: integer
      value:
        type: string
    type: object
  domain.EsProductRelatedInfo:
    properties:
      brandNames:
        items:
          type: string
        type: array
      productAttrs:
        items:
          $ref: '#/definitions/domain.ProductAttr'
        type: array
      productCategoryNames:
        items:
          type: string
        type: array
    type: object
  domain.MemberReadHistory:
    properties:
      createTime:
        type: string
      id:
        type: string
      memberIcon:
        type: string
      memberId:
        type: integer
      memberNick:
        type: string
      productId:
        type: integer
      productName:
        type: string
      productPic:
        type: string
      productPrice:
        type: string
      productSubTittle:
        type: string
    type: object
  domain.ProductAttr:
    properties:
      attrId:
        type: integer
      attrName:
        type: string
      attrValues:
        items:
          type: string
        type: array
    type: object
...
  response.ResponseMsg:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
  description: mall
  termsOfService: https://github.com/18211167516/Go-Gin-Api
  title: mall
  version: "1.0"
paths:
...
  /esProduct/create/{id}:
    post:
      consumes:
      - application/json
      description: 根据id创建商品
      operationId: v1/esProduct/Create
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/document.EsProduct'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 根据id创建商品
      tags:
      - 搜索商品管理
  /esProduct/delete/{id}:
    get:
      consumes:
      - application/json
      description: 根据id删除商品
      operationId: v1/esProduct/Delete
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      summary: 根据id删除商品
      tags:
      - 搜索商品管理
  /esProduct/delete/batch:
    post:
      consumes:
      - application/json
      description: 根据id批量删除商品
      operationId: v1/esProduct/DeleteBatch
      parameters:
      - description: product_ids
        in: query
        items:
          type: integer
        name: ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 根据id批量删除商品
      tags:
      - 搜索商品管理
  /esProduct/importAll:
    post:
      consumes:
      - application/json
      description: 导入所有数据库中商品到ES
      operationId: v1/esProduct/importAll
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      summary: 导入所有数据库中商品到ES
      tags:
      - 搜索商品管理
  /esProduct/recommend/{id}:
    get:
      consumes:
      - application/json
      description: 根据商品id推荐商品
      operationId: v1/esProduct/Recommend
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - default: 0
        description: page number
        in: query
        name: pageNum
        type: integer
      - default: 5
        description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: default
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 根据商品id推荐商品
      tags:
      - 搜索商品管理
  /esProduct/search:
    get:
      consumes:
      - application/json
      description: 综合搜索、筛选、排序
      operationId: v1/esProduct/SearchDetail
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      - default: 0
        description: brandId
        in: query
        name: brandId
        type: integer
      - default: 0
        description: product Category Id
        in: query
        name: productCategoryId
        type: integer
      - default: 0
        description: page number
        in: query
        name: pageNum
        type: integer
      - default: 5
        description: page size
        in: query
        name: pageSize
        type: integer
      - description: sort 排序字段:0->按相关度；1->按新品；2->按销量；3->价格从低到高；4->价格从高到低
        in: query
        name: sort
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 综合搜索、筛选、排序
      tags:
      - 搜索商品管理
  /esProduct/search/relate:
    get:
      consumes:
      - application/json
      description: 获取搜索的相关品牌、分类及筛选属性
      operationId: v1/esProduct/SearchRelatedInfo
      parameters:
      - description: keyword
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/domain.EsProductRelatedInfo'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 获取搜索的相关品牌、分类及筛选属性
      tags:
      - 搜索商品管理
  /esProduct/search/simple:
    get:
      consumes:
      - application/json
      description: 简单搜索
      operationId: v1/esProduct/SearchSimple
      parameters:
      - default: '""'
        description: page number
        in: query
        name: keyword
        type: string
      - default: 0
        description: page number
        in: query
        name: pageNum
        type: integer
      - default: 5
        description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.ResponseMsg'
            - properties:
                data:
                  $ref: '#/definitions/document.EsProduct'
              type: object
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 简单搜索
      tags:
      - 搜索商品管理
  /member/readHistory/clear:
    post:
      consumes:
      - application/json
      description: 清空浏览记录
      operationId: v1/MemberReadHistoryController/Clear
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 清空浏览记录
      tags:
      - 用户浏览记录接口
  /member/readHistory/create:
    post:
      consumes:
      - application/json
      description: 创建一个浏览记录
      operationId: v1/MemberReadHistoryController/Create
      parameters:
      - description: MemberReadHistory
        in: body
        name: MemberReadHistory
        required: true
        schema:
          $ref: '#/definitions/domain.MemberReadHistory'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 创建浏览记录
      tags:
      - 用户浏览记录接口
  /member/readHistory/delete:
    post:
      consumes:
      - application/json
      description: 根据ids清空出浏览记录
      operationId: v1/MemberReadHistoryController/Delete
      parameters:
      - description: history_ids
        in: query
        items:
          type: string
        name: ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 删除浏览记录
      tags:
      - 用户浏览记录接口
  /member/readHistory/list:
    get:
      consumes:
      - application/json
      description: 分页获取浏览记录列表
      operationId: v1/MemberReadHistoryController/List
      parameters:
      - default: 0
        description: page number
        in: query
        name: pageNum
        type: integer
      - default: 5
        description: page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/response.ResponseMsg'
        "500":
          description: failure
          schema:
            $ref: '#/definitions/response.ResponseMsg'
      security:
      - JWT: []
      summary: 获取浏览记录列表
      tags:
      - 用户浏览记录接口
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"